requirements.txt
attrs
beautifulsoup4
phonenumbers
requests
